requirements_python2.txt
PyYAML==3.11
beautifulsoup4==4.4.0
requests==2.7.0
wsgiref==0.1.2
